fix: Addressing review comments
